+++ picocli-annotation-processing-tests/src/test/resources/picocli/examples/validation/Invalid2.java
@@ -1,0 +1,9 @@
+package picocli.examples.validation;
+
+import static picocli.CommandLine.Option;
+
+public class Invalid2 {
+
+    @Option(names = "--negatable", negatable = true)
+    int invalidNegatableShouldBeBoolean;
+}